include/includes.h: lib/system.c: Can't assume every system has a statvfs varient.               Return -1 for those that don't. smbd/reply.c: Fixed printf warning. Jeremy. (This used to be commit 14c134e8316687aa5a4ee089c2acfa6428faceae)

+++ source3/smbd/reply.c
@@ -932,7 +932,7 @@ int reply_dskattr(connection_struct *conn, char *inbuf,char *outbuf, int dum_siz
   SSVAL(outbuf,smb_vwv2,512);
   SSVAL(outbuf,smb_vwv3,dfree);
 
-  DEBUG(3,("dskattr dfree=%d\n", dfree));
+  DEBUG(3,("dskattr dfree=%d\n", (unsigned int)dfree));
 
   return(outsize);
 }
